+++ 1.21.10/Common/src/main/java/fuzs/statuemenus/api/v1/client/gui/components/FlatTickButton.java
@@ -1,20 +1,20 @@
 package fuzs.statuemenus.api.v1.client.gui.components;
 
 import fuzs.statuemenus.api.v1.client.gui.screens.AbstractStatueScreen;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.input.InputWithModifiers;
 import net.minecraft.client.renderer.RenderPipelines;
-import net.minecraft.client.renderer.texture.Tickable;
 import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.ARGB;
 
-public class FlatTickButton extends FlatButton implements Tickable {
+public class FlatTickButton extends FlatButton {
     private final int imageTextureX;
     private final int imageTextureY;
     private final ResourceLocation imageTextureLocation;
-    private int lastClickedTicks;
+    private long lastClickedTime;
     protected int lastClickedTicksDelay = 20;
 
     public FlatTickButton(int x, int y, int width, int height, int imageTextureX, int imageTextureY, ResourceLocation imageTextureLocation, OnPress onPress) {
@@ -46,21 +46,19 @@ public class FlatTickButton extends FlatButton implements Tickable {
     @Override
     public void onPress(InputWithModifiers inputWithModifiers) {
         super.onPress(inputWithModifiers);
-        this.lastClickedTicks = this.lastClickedTicksDelay;
+        this.lastClickedTime = Util.getMillis();
     }
 
-    @Override
-    public void tick() {
-        if (this.lastClickedTicks > 0) this.lastClickedTicks--;
+    protected boolean wasClicked() {
+        return Util.getMillis() - this.lastClickedTime < this.lastClickedTicksDelay * 50L;
     }
 
     @Override
     protected void renderBg(GuiGraphics guiGraphics, Minecraft minecraft, int mouseX, int mouseY) {
         super.renderBg(guiGraphics, minecraft, mouseX, mouseY);
-        ResourceLocation resourceLocation =
-                this.lastClickedTicks > 0 ? this.textureLocation : this.imageTextureLocation;
+        ResourceLocation resourceLocation = this.wasClicked() ? this.textureLocation : this.imageTextureLocation;
         final int i = this.getYImage();
-        if (this.lastClickedTicks > 0) {
+        if (this.wasClicked()) {
             guiGraphics.blit(RenderPipelines.GUI_TEXTURED,
                     resourceLocation,
                     this.getX() + this.width / 2 - 8,