+++ 1.21.10/Common/src/main/java/fuzs/statuemenus/api/v1/client/gui/components/TickButton.java
@@ -1,18 +1,18 @@
 package fuzs.statuemenus.api.v1.client.gui.components;
 
 import fuzs.statuemenus.api.v1.client.gui.screens.AbstractStatueScreen;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.input.InputWithModifiers;
 import net.minecraft.client.renderer.RenderPipelines;
-import net.minecraft.client.renderer.texture.Tickable;
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.ARGB;
 
-public class TickButton extends FlatButton implements Tickable {
+public class TickButton extends FlatButton {
     private final Component title;
     private final Component clickedTitle;
-    private int lastClickedTicks;
+    private long lastClickedTime;
     protected int lastClickedTicksDelay = 30;
 
     public TickButton(int x, int y, int width, int height, Component title, Component clickedTitle, OnPress onPress) {
@@ -24,22 +24,19 @@ public class TickButton extends FlatButton implements Tickable {
     @Override
     public void onPress(InputWithModifiers inputWithModifiers) {
         super.onPress(inputWithModifiers);
-        this.lastClickedTicks = this.lastClickedTicksDelay;
+        this.lastClickedTime = Util.getMillis();
         this.setMessage(this.clickedTitle);
     }
 
-    @Override
-    public void tick() {
-        if (this.lastClickedTicks > 0) this.lastClickedTicks--;
-        if (this.lastClickedTicks == 0) {
-            this.setMessage(this.title);
-        }
+    protected boolean wasClicked() {
+        return Util.getMillis() - this.lastClickedTime < this.lastClickedTicksDelay * 50L;
     }
 
     @Override
     protected void renderBg(GuiGraphics guiGraphics, Minecraft minecraft, int mouseX, int mouseY) {
         super.renderBg(guiGraphics, minecraft, mouseX, mouseY);
-        if (this.lastClickedTicks > 0) {
+        if (this.wasClicked()) {
+            this.setMessage(this.clickedTitle);
             final int i = this.getYImage();
             int titleWidth = minecraft.font.width(this.clickedTitle);
             final int startX = (this.width - titleWidth - (titleWidth > 0 ? 4 : 0) - 16) / 2;
@@ -54,14 +51,17 @@ public class TickButton extends FlatButton implements Tickable {
                     256,
                     256,
                     ARGB.white(this.alpha));
+        } else {
+            this.setMessage(this.title);
         }
     }
 
     @Override
     protected int getMessageXOffset() {
-        if (this.lastClickedTicks > 0) {
+        if (this.wasClicked()) {
             return (16 + 4) / 2;
+        } else {
+            return super.getMessageXOffset();
         }
-        return super.getMessageXOffset();
     }
 }