+++ 1.21.10/Common/src/main/java/fuzs/statuemenus/api/v1/client/gui/screens/StatueTickBoxScreen.java
@@ -33,6 +33,7 @@ public abstract class StatueTickBoxScreen<T> extends AbstractStatueScreen {
         if (this.inputUpdateTicks > 0) {
             this.inputUpdateTicks--;
         }
+
         this.testNameInputChanged(true);
     }