+++ 1.21.10/CHANGELOG.md
@@ -5,13 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v21.10.3-1.21.10] - 2025-10-14
+## [v21.10.5-1.21.10] - 2025-10-15
+
+### Changed
+
+- New reset button for some sliders
+- Tick buttons no longer depend on game ticks
+
+## [v21.10.4-1.21.10] - 2025-10-14
 
 ### Changed
 
 - Rework parts of `StatueWidgetsScreen`
 
-## [v21.10.4-1.21.10] - 2025-10-14
+## [v21.10.3-1.21.10] - 2025-10-14
 
 ### Changed